gebug, bit utility -> Getools.Lib
