+++ Getools/Gebug64.Unfloader/Message/CommandParameter/U8Parameter.cs
@@ -1,9 +1,10 @@
-﻿using Getools.Lib.Architecture;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Getools.Lib;
+using Getools.Lib.Architecture;
 
 namespace Gebug64.Unfloader.Message.CommandParameter
 {
