+++ Getools/Gebug64.Unfloader/Message/RomAckMessage.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Gebug64.Unfloader.Message.MessageType;
 using Gebug64.Unfloader.UsbPacket;
+using Getools.Lib;
 
 namespace Gebug64.Unfloader.Message
 {
