+++ Getools/Getools.Lib/BitUtility.cs
@@ -246,6 +246,54 @@ namespace Getools.Lib
         }
 
         /// <summary>
+        /// Converts 16 bit value to LSB and inserts into array at index.
+        /// </summary>
+        /// <param name="arr">Array to insert value into.</param>
+        /// <param name="index">Index to insert value at.</param>
+        /// <param name="value">Value to insert.</param>
+        public static void Insert16Little(byte[] arr, Int16 index, Int16 value)
+        {
+            arr[index + 0] = (byte)(value & 0xff);
+            arr[index + 1] = (byte)((value >> 8) & 0xff);
+        }
+
+        /// <summary>
+        /// Converts 16 bit value to LSB and inserts into array at index.
+        /// </summary>
+        /// <param name="arr">Array to insert value into.</param>
+        /// <param name="index">Index to insert value at.</param>
+        /// <param name="value">Value to insert.</param>
+        public static void Insert16Little(byte[] arr, Int16 index, UInt16 value)
+        {
+            arr[index + 0] = (byte)(value & 0xff);
+            arr[index + 1] = (byte)((value >> 8) & 0xff);
+        }
+
+        /// <summary>
+        /// Converts 16 bit value to MSB and inserts into array at index.
+        /// </summary>
+        /// <param name="arr">Array to insert value into.</param>
+        /// <param name="index">Index to insert value at.</param>
+        /// <param name="value">Value to insert.</param>
+        public static void Insert16Big(byte[] arr, Int16 index, Int16 value)
+        {
+            arr[index + 0] = (byte)((value >> 8) & 0xff);
+            arr[index + 1] = (byte)(value & 0xff);
+        }
+
+        /// <summary>
+        /// Converts 16 bit value to MSB and inserts into array at index.
+        /// </summary>
+        /// <param name="arr">Array to insert value into.</param>
+        /// <param name="index">Index to insert value at.</param>
+        /// <param name="value">Value to insert.</param>
+        public static void Insert16Big(byte[] arr, Int16 index, UInt16 value)
+        {
+            arr[index + 0] = (byte)((value >> 8) & 0xff);
+            arr[index + 1] = (byte)(value & 0xff);
+        }
+
+        /// <summary>
         /// Converts 32 bit value to LSB and inserts into array at index.
         /// </summary>
         /// <param name="arr">Array to insert value into.</param>
@@ -313,6 +361,31 @@ namespace Getools.Lib
             s |= (short)stream.ReadByte();
 
             return s;
+        }
+
+        /// <summary>
+        /// Reads 2 bytes from a byte array as big endian 16-bit int.
+        /// </summary>
+        /// <param name="arr">Array to read.</param>
+        /// <param name="index">Starting index in array.</param>
+        /// <returns>Int.</returns>
+        public static Int16 Read16Big(byte[] arr, int index)
+        {
+            if (index < 0)
+            {
+                throw new ArgumentException("Array index must be non-negative integer");
+            }
+
+            if (index + 2 > arr.Length)
+            {
+                throw new EndOfStreamException("Reading 4 bytes from array exceeds array length");
+            }
+
+            Int16 i = arr[index];
+            i <<= 8;
+            i |= arr[index + 1];
+
+            return i;
         }
 
         /// <summary>
@@ -588,6 +661,24 @@ namespace Getools.Lib
             }
 
             return sb.ToString();
+        }
+
+        /// <summary>
+        /// UNFLoader calc_padsize.
+        /// </summary>
+        /// <param name="size"></param>
+        /// <returns></returns>
+        public static UInt32 CalculatePadsize(UInt32 size)
+        {
+            size--;
+            size |= size >> 1;
+            size |= size >> 2;
+            size |= size >> 4;
+            size |= size >> 8;
+            size |= size >> 16;
+            size++;
+
+            return size;
         }
     }
 }
